requirements.txt
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cycler==0.11.0
fonttools==4.37.1
imageio==2.21.2
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
tifffile==2022.8.12
wincertstore==0.2
